Make traces MUCH more efficient
Changed the underlying representation of a trace to a type that
naturally compresses consecutive runs of pages. This decreases the
memory size of serialized Trace objects by like a factor of 10.

Also, made working with Traces a lot faster by slimming down the number
of types used and taking advantage of the sorted nature of the data.

Plots actually generate in a reasonable amount of time now.